requirements.txt
google-generativeai
google-cloud-speech

requests
SpeechRecognition
pyaudio
pygame
websockets
beautifulsoup4
